+++ apps/vmq_diversity/src/vmq_diversity.erl
@@ -27,13 +27,13 @@ start() ->
 load_script(Script) when is_list(Script) ->
     case filelib:is_file(Script) of
         true ->
-            vmq_diversity_script_sup:start_script(Script);
+            vmq_diversity_scripts_sup:start_script(Script);
         false ->
             {error, cant_find_script}
     end.
 
 reload_script(Script) when is_list(Script) ->
-    vmq_diversity_script_sup:reload_script(Script).
+    vmq_diversity_scripts_sup:reload_script(Script).
 
 unload_script(Script) when is_list(Script) ->
-    vmq_diversity_script_sup:stop_script(Script).
+    vmq_diversity_scripts_sup:stop_script(Script).